Added a new package parameter to import the Java windows service config Updated documentation Cleaned up the install script. Removing variables not needed, code not executed etc. Added Checksums to downloaded files Changed the NEO4J_HOME detection in uninstall to be smarter Removes the NEO4J_HOME env variable on uninstall.

+++ neo4j-community/tools/chocolateyUninstall.ps1
@@ -1,18 +1,28 @@
 $PackageName = 'neo4j-community'
 try {
-
-  $neoHome = $Env:NEO4J_HOME
+  # Get the NeoHome Dir
+  #   Try the local environment
+  $neoHome = [string] (Get-EnvironmentVariable -Name 'NEO4J_HOME' -Scope 'Machine')  
+  # Failing that, try a registry hack
+  if ($neoHome -eq '')
+  {
+    $neoHome = [string] ( (Get-ItemProperty -Path 'Registry::HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Control\Session Manager\Environment' -ErrorAction Ignore).'NEO4J_HOME' )
+  }
+  if ($neoHome -eq '') { throw "Could not find the Neo4jHome directory" }
 
   # Uninstall the Neo4j Service
   $UninstallBatch = "$($neoHome)\bin\Neo4jInstaller.bat"
   if (!(Test-Path $UninstallBatch)) { throw "Could not find the Neo4j Uninstaller Batch file at $UninstallBatch" }
   
-  # Need to use a new environment as the NEO4J_HOME may not have been set correctly
+  # Uninstall the service
   $args = "remove"
   $results = (Start-Process -FilePath $UninstallBatch -ArgumentList $args -Wait -PassThru -NoNewWindow -UseNewEnvironment)
 
   # Remove the install folder
   Remove-Item -Path $neoHome -Recurse -Force | Out-Null
+  
+  # Remove the environment variable
+  Install-ChocolateyEnvironmentVariable "NEO4J_HOME" '' "Machine"
 
   Write-ChocolateySuccess $PackageName
 } catch {
